Add multi-epoch and pipeline throttling support to shufflers.
